+++ pkg/installer/installer.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/pkg/config/remote/client"
-	installerErrors "github.com/DataDog/datadog-agent/pkg/installer/errors"
+	installerErrors "github.com/DataDog/datadog-agent/pkg/installer/packages/errors"
 	"github.com/DataDog/datadog-agent/pkg/installer/packages/repository"
 	"github.com/DataDog/datadog-agent/pkg/installer/packages/service"
 	pbgo "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
