+++ pkg/installer/packages/errors/errors_test.go
